+++ lib/platform/android/android_data_loader.cpp
@@ -1,6 +1,6 @@
 #include <android/asset_manager.h>
 #include <android_native_app_glue.h>
-#include <platform/android/android_data_store.hpp>
+#include <platform/android/android_data_loader.hpp>
 #include <memory>
 
 namespace bave {
@@ -28,9 +28,13 @@ auto do_read_data(android_app* app, Type& out, CString path) -> bool {
 }
 } // namespace
 
-auto AndroidDataStore::do_exists(CString path) const -> bool { return open_asset(m_app->activity->assetManager, path) != nullptr; }
+auto AndroidDataLoader::exists(std::string_view const uri) const -> bool {
+	return open_asset(m_app->activity->assetManager, std::string{uri}.c_str()) != nullptr;
+}
 
-auto AndroidDataStore::do_read_bytes(std::vector<std::byte>& out, CString path) const -> bool { return do_read_data(m_app, out, path); }
+auto AndroidDataLoader::read_bytes(std::vector<std::byte>& out, std::string_view const uri) const -> bool {
+	return do_read_data(m_app, out, std::string{uri}.c_str());
+}
 
-auto AndroidDataStore::do_read_string(std::string& out, CString path) const -> bool { return do_read_data(m_app, out, path); }
+auto AndroidDataLoader::read_string(std::string& out, std::string_view const uri) const -> bool { return do_read_data(m_app, out, std::string{uri}.c_str()); }
 } // namespace bave