+++ lib/src/data_store.cpp
@@ -1,5 +1,6 @@
 #include <bave/data_store.hpp>
 #include <djson/json.hpp>
+#include <algorithm>
 #include <filesystem>
 
 namespace bave {
@@ -16,39 +17,45 @@ auto DataStore::as_string_view(std::span<std::byte const> bytes) -> std::string_
 	return {reinterpret_cast<char const*>(bytes.data()), bytes.size()}; // NOLINT(cppcoreguidelines-pro-type-reinterpret-cast)
 }
 
+void DataStore::add_loader(std::unique_ptr<IDataLoader> loader, int const priority) {
+	if (!loader) { return; }
+	auto const it = std::lower_bound(m_loaders.begin(), m_loaders.end(), priority, [](Entry const& a, int const priority) { return a.priority < priority; });
+	auto entry = Entry{.loader = std::move(loader), .priority = priority};
+	m_loaders.insert(it, std::move(entry));
+}
+
+auto DataStore::exists(std::string_view uri) const -> bool {
+	if (uri.empty()) { return false; }
+	for (auto const& [loader, _] : m_loaders) {
+		if (loader->exists(uri)) { return true; }
+	}
+	return false;
+}
+
 auto DataStore::read_bytes(std::string_view uri) const -> std::vector<std::byte> {
 	if (uri.empty()) { return {}; }
-	auto path = make_full_path(uri);
 	auto ret = std::vector<std::byte>{};
-	if (!do_read_bytes(ret, path.c_str())) {
-		m_log.warn("failed to load data at: '{}'", uri);
-		return {};
+	for (auto const& [loader, _] : m_loaders) {
+		if (loader->read_bytes(ret, uri)) { return ret; }
 	}
-	return ret;
+	m_log.warn("failed to load data at: '{}'", uri);
+	return {};
 }
 
 auto DataStore::read_string(std::string_view uri) const -> std::string {
 	if (uri.empty()) { return {}; }
-	auto path = make_full_path(uri);
 	auto ret = std::string{};
-	if (!do_read_string(ret, path.c_str())) {
-		m_log.warn("failed to load data at: '{}'", uri);
-		return {};
+	for (auto const& [loader, _] : m_loaders) {
+		if (loader->read_string(ret, uri)) { return ret; }
 	}
-	return ret;
+	m_log.warn("failed to load data at: '{}'", uri);
+	return {};
 }
 
 auto DataStore::read_json(std::string_view uri) const -> dj::Json {
 	auto str = read_string(uri);
 	if (str.empty()) { return {}; }
 	return dj::Json::parse(str);
-}
-
-auto DataStore::make_full_path(std::string_view uri) const -> std::string { return (fs::path{m_prefix} / uri).generic_string(); }
-
-auto DataStore::make_uri(std::string_view const full_path) const -> std::string {
-	if (full_path.empty()) { return {}; }
-	return fs::path{full_path}.lexically_relative(get_mount_point()).generic_string();
 }
 
 auto DataStore::to_spir_v(std::string_view const glsl) const -> std::string {
@@ -58,12 +65,5 @@ auto DataStore::to_spir_v(std::string_view const glsl) const -> std::string {
 		return {};
 	}
 	return spir_v_uri;
-}
-
-auto DataStore::do_read_string(std::string& out, CString path) const -> bool {
-	auto bytes = std::vector<std::byte>{};
-	if (!do_read_bytes(bytes, path)) { return false; }
-	out = std::string{reinterpret_cast<char const*>(bytes.data()), bytes.size()}; // NOLINT(cppcoreguidelines-pro-type-reinterpret-cast)
-	return true;
 }
 } // namespace bave