+++ lib/docs/changelog.md
@@ -4,7 +4,7 @@
 
 ### v0.4.10
 
-- bave::DataStore now stores an ordered list of bave::DataLoader s, users can add their own loaders to the list.
+- bave::DataStore now stores an ordered list of bave::IDataLoader s, users can add their own loaders to the list.
 
 ### v0.4.9